requirements.txt
fastapi==0.110.0
uvicorn==0.29.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# LLM (ChatGroq)
groq==0.9.0

# Environment variables
python-dotenv==1.0.1

# Validation & schemas
pydantic==2.6.4
